requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
cycler==0.10.0
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
